ai-vision-service/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
torch==2.3.1
torchvision==0.18.1
pydantic==2.8.2
openai==1.47.0
starlette==0.37.2
